added tests, changed main.js, added MainRoot.jsx from previous main.js
