+++ package.json
@@ -13,7 +13,8 @@
     "format": "prettier --write 'src/**/*.{js,jsx,css,html}'",
     "preview": "vite preview",
     "start:both": "concurrently \"vite\" \"cd build/target && java -jar server-1.1.1-SNAPSHOT.jar\"",
-    "start:both-prod": "concurrently \"vite dist --port 3030 --host 0.0.0.0\" \"cd target && java -jar server-1.1.1-SNAPSHOT.jar\""
+    "start:both-prod": "concurrently \"vite dist --port 3030 --host 0.0.0.0\" \"cd target && java -jar server-1.1.1-SNAPSHOT.jar\"",
+    "test": "vitest"
   },
   "dependencies": {
     "@stomp/stompjs": "^7.0.0",
@@ -25,7 +26,6 @@
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.2",
     "react-router": "^6.15.0",
-    "react-router-dom": "^6.15.0",
     "ws": "^8.13.0"
   },
   "devDependencies": {
@@ -36,6 +36,9 @@
     "@babel/register": "^7.22.5",
     "@reduxjs/toolkit": "^1.9.5",
     "@stomp/stompjs": "^7.0.0",
+    "@tanstack/react-query": "^5.8.4",
+    "@testing-library/jest-dom": "^6.1.4",
+    "@testing-library/react": "^13.4.0",
     "@vitejs/plugin-react": "^4.0.3",
     "autoprefixer": "^10.4.16",
     "babel-loader": "^9.1.3",
@@ -46,13 +49,16 @@
     "eslint": "^8.45.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
+    "global-jsdom": "^9.1.0",
+    "happy-dom": "^7.6.0",
+    "jsdom": "^22.1.0",
     "net": "^1.0.2",
     "nodemon": "^3.0.1",
     "postcss": "^8.4.31",
     "react-chartjs-2": "^5.2.0",
     "react-redux": "^8.1.3",
     "react-router": "^6.15.0",
-    "react-router-dom": "^6.15.0",
+    "react-router-dom": "^6.4.1",
     "redux-toolkit": "^1.1.2",
     "sass": "^1.65.1",
     "sass-loader": "^13.3.2",
@@ -62,7 +68,8 @@
     "stream-http": "^3.2.0",
     "style-loader": "^3.3.3",
     "tailwindcss": "^3.3.5",
-    "vite": "^4.4.5"
+    "vite": "^4.4.5",
+    "vitest": "^0.34.6",
+    "vitest-fetch-mock": "^0.2.2"
   }
 }
-
