+++ root.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+
+import MainRoot from './src/MainRoot';
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+root.render(<MainRoot />);
