+++ src/containers/About.jsx
@@ -3,7 +3,12 @@ import 'font-awesome/css/font-awesome.min.css';
 const About = () => {
   return (
     <div className="settings-container">
-      <img style={{ width: '250px' }} src="src/assets/logo.png" alt="" />
+      <img
+        data-testid="kaffee-logo"
+        style={{ width: '250px' }}
+        src="src/assets/logo.png"
+        alt=""
+      />
 
       <p>v1.1.1</p>
       <br />
@@ -11,7 +16,11 @@ const About = () => {
         Kaffee is a free and open source product <br /> in early development.
       </p>
       <p>To contribute, please open an issue or pull request.</p>
-      <a href="https://github.com/oslabs-beta/Kaffee" target="_blank">
+      <a
+        data-testid="github-link"
+        href="https://github.com/oslabs-beta/Kaffee"
+        target="_blank"
+      >
         <i className="fa fa-github" style={{ fontSize: '36px' }}></i>
       </a>
     </div>
